Move components to folders to keep everything organized
