+++ src/components/index.ts
@@ -1,1 +1,3 @@
-export { Sidebar } from "./Sidebar";
+export { Sidebar } from "./sidebar/Sidebar";
+export { SimpleWidget } from "./dashboard/SimpleWidget";
+export { WidgetsGrid } from "./dashboard/WidgetsGrid";
